@@ -1,10 +1,10 @@
 name: flutter_colorpicker
 description: HSV(HSB)/HSL/RGB/Material color picker inspired by all the good design for your amazing flutter apps.
-version: 1.0.3
+version: 1.1.0
 homepage: https://github.com/mchome/flutter_colorpicker
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.14.0 <4.0.0"
   flutter: ">=2.0.0"
 
 dependencies:
@@ -14,4 +14,4 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
+  flutter_lints: ^4.0.0
