@@ -1,6 +1,6 @@
-/// The components of HSV Color Picker
-///
-/// Try to create a Color Picker with other layout on your own :)
+// The components of HSV Color Picker
+//
+// Try to create a Color Picker with other layout on your own :)
 
 import 'dart:math';
 import 'package:flutter/gestures.dart';
@@ -79,6 +79,7 @@ class HSVWithHueColorPainter extends CustomPainter {
       Paint()
         ..color = pointerColor ?? (useWhiteForeground(hsvColor.toColor()) ? Colors.white : Colors.black)
         ..strokeWidth = 1.5
+        ..blendMode = BlendMode.luminosity
         ..style = PaintingStyle.stroke,
     );
   }
@@ -824,7 +825,7 @@ class ColorPickerLabel extends StatefulWidget {
   final List<ColorLabelType> colorLabelTypes;
 
   @override
-  _ColorPickerLabelState createState() => _ColorPickerLabelState();
+  State<ColorPickerLabel> createState() => _ColorPickerLabelState();
 }
 
 class _ColorPickerLabelState extends State<ColorPickerLabel> {
@@ -896,14 +897,14 @@ class _ColorPickerLabelState extends State<ColorPickerLabel> {
                   children: <Widget>[
                     Text(
                       item,
-                      style: widget.textStyle ?? Theme.of(context).textTheme.bodyText1,
+                      style: widget.textStyle ?? Theme.of(context).textTheme.bodyLarge,
                     ),
                     const SizedBox(height: 10.0),
                     Expanded(
                       child: Text(
                         colorValue(widget.hsvColor, _colorType)[_colorTypes[_colorType]!.indexOf(item)],
                         overflow: TextOverflow.ellipsis,
-                        style: widget.textStyle ?? Theme.of(context).textTheme.bodyText2,
+                        style: widget.textStyle ?? Theme.of(context).textTheme.bodyMedium,
                       ),
                     ),
                   ],
@@ -954,7 +955,7 @@ class ColorPickerInput extends StatefulWidget {
   final bool disable;
 
   @override
-  _ColorPickerInputState createState() => _ColorPickerInputState();
+  State<ColorPickerInput> createState() => _ColorPickerInputState();
 }
 
 class _ColorPickerInputState extends State<ColorPickerInput> {
@@ -970,6 +971,7 @@ class _ColorPickerInputState extends State<ColorPickerInput> {
   @override
   Widget build(BuildContext context) {
     if (inputColor != widget.color.value) {
+      // ignore: prefer_interpolation_to_compose_strings
       textEditingController.text = '#' +
           widget.color.red.toRadixString(16).toUpperCase().padLeft(2, '0') +
           widget.color.green.toRadixString(16).toUpperCase().padLeft(2, '0') +
@@ -979,10 +981,10 @@ class _ColorPickerInputState extends State<ColorPickerInput> {
     return Padding(
       padding: const EdgeInsets.only(top: 5.0),
       child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-        if (!widget.embeddedText) Text('Hex', style: Theme.of(context).textTheme.bodyText1),
+        if (!widget.embeddedText) Text('Hex', style: Theme.of(context).textTheme.bodyLarge),
         const SizedBox(width: 10),
         SizedBox(
-          width: (Theme.of(context).textTheme.bodyText2?.fontSize ?? 14) * 10,
+          width: (Theme.of(context).textTheme.bodyMedium?.fontSize ?? 14) * 10,
           child: TextField(
             enabled: !widget.disable,
             controller: textEditingController,
