@@ -133,7 +133,7 @@ Widget pickerLayoutBuilder(BuildContext context, List<Color> colors, PickerItem 
             },
             child: Icon(Icons.code, color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
             style: ElevatedButton.styleFrom(
-              primary: widget.pickerColor,
+              backgroundColor: widget.pickerColor,
               shadowColor: widget.pickerColor.withOpacity(1),
               elevation: 10,
             ),
@@ -218,7 +218,7 @@ Widget pickerItemBuilder(Color color, bool isCurrentColor, void Function() chang
             },
             child: Icon(Icons.code, color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
             style: ElevatedButton.styleFrom(
-              primary: widget.pickerColor,
+              backgroundColor: widget.pickerColor,
               shadowColor: widget.pickerColor.withOpacity(1),
               elevation: 10,
             ),
@@ -300,7 +300,7 @@ Widget pickerItemBuilder(Color color, bool isCurrentColor, void Function() chang
                 style: TextStyle(color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
               ),
               style: ElevatedButton.styleFrom(
-                primary: widget.pickerColor,
+                backgroundColor: widget.pickerColor,
                 shadowColor: widget.pickerColor.withOpacity(1),
                 elevation: 10,
               ),
@@ -333,7 +333,7 @@ BlockPicker(
               },
               child: Icon(Icons.code, color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
               style: ElevatedButton.styleFrom(
-                primary: widget.pickerColor,
+                backgroundColor: widget.pickerColor,
                 shadowColor: widget.pickerColor.withOpacity(1),
                 elevation: 10,
               ),
@@ -372,7 +372,7 @@ BlockPicker(
                 style: TextStyle(color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
               ),
               style: ElevatedButton.styleFrom(
-                primary: widget.pickerColor,
+                backgroundColor: widget.pickerColor,
                 shadowColor: widget.pickerColor.withOpacity(1),
                 elevation: 10,
               ),
@@ -405,7 +405,7 @@ MultipleChoiceBlockPicker(
               },
               child: Icon(Icons.code, color: useWhiteForeground(widget.pickerColor) ? Colors.white : Colors.black),
               style: ElevatedButton.styleFrom(
-                primary: widget.pickerColor,
+                backgroundColor: widget.pickerColor,
                 shadowColor: widget.pickerColor.withOpacity(1),
                 elevation: 10,
               ),
